* still flip/flopping on the index type. Started with `uintpr_t`, then moved to `size_t` and now `unsigned int`. Don't want to typedef this though. See dox/DATATYPES.md for this. * moved prime and hash code to new project mulle-data * removed some stub documentation, that now does more harm than good * added verb modifiers to the function naming style doc * added `describe` for debugging * `guarantee` now returns a non-null pointer on success * new range get methods like `_mulle__array_get_in_range` for example * `compact` as a verb is no more, use `size_to_fit` * enumerators now nil the by reference item when they are exhausted fixed a few missing inline declarations

+++ src/rangeset/mulle--rangeset.h
@@ -27,14 +27,14 @@
 struct mulle__rangeset
 {
    struct mulle_range   *_ranges;
-   size_t                _length;
-   size_t                _size;
+   unsigned int         _length;
+   unsigned int         _size;
 };
 
 
 MULLE_C_NONNULL_FIRST
 static inline void   _mulle__rangeset_init( struct mulle__rangeset *p,
-                                            size_t capacity,
+                                            unsigned int capacity,
                                             struct mulle_allocator *allocator)
 {
    memset( p, 0, sizeof( *p));
